Bug-Fix: false positive for empty root directory
If a root directory does not contain anything at all then the watch dog
would comply that a package with name "" (the default package) does not
match any naming rule (if there is none matching "" of course). This is
not the intended behavior, an empty root directory can safely be
ignored.